Finder refactoring   - Renamed blueflood-graphite-finder to blueflood_graphite_finder as per pep standards   - Made blueflood_graphite_finder a package   - Added logging config file for tests   - Fixed setup.py file
