+++ .gitignore
@@ -81,5 +81,5 @@ blueflood_graphite_finder.egg-info/*
 contrib/enum_metrics/dist/
 
 # Python egg metadata, regenerated from source files by setuptools.
-contrib/enum_metrics/dist/*.egg-info
-contrib/enum_metrics/dist/*.egg
+*.egg-info
+*.egg
