+++ setup.py
@@ -1,5 +1,5 @@
 # coding: utf-8
-from setuptools import setup
+from setuptools import setup, find_packages
 
 version = '0.0.3'
