+++ tests/logging.ini
@@ -1,0 +1,28 @@
+[loggers]
+keys=root,blueflood_finder
+
+[handlers]
+keys=consoleHandler
+
+[formatters]
+keys=simpleFormatter
+
+[logger_root]
+level=DEBUG
+handlers=consoleHandler
+
+[logger_blueflood_finder]
+level=INFO
+handlers=consoleHandler
+qualname=blueflood_finder
+propagate=0
+
+[handler_consoleHandler]
+class=StreamHandler
+level=DEBUG
+formatter=simpleFormatter
+args=(sys.stdout,)
+
+[formatter_simpleFormatter]
+format=%(asctime)s - %(name)s - %(levelname)s - %(message)s
+datefmt=
